Crate aoc_2021_day_15

Advent of Code 2021: Day 15: Chiton

https://adventofcode.com/2021/day/15

Input

A grid of single-digit cost values, in the range 1 to 9.

Part one

Calculate the lower cost to get from the top left point to the bottom right. The cost is applied on entering a square, so the top left cost is irrelevant.

Part two

Same as part one, except the grid is replicated five times down and to the right, for a total of twenty five times. Each time the grid is replicated in those directions, the cost of each point is increased by one. A value of 9 becomes 1.

Algorithm

This is a relatively simple path finder. We use PrioritySearch to prioritise the routes with the lowest risk, and keep a cache of the best ones so far to each point so we can short-circuit appropriately.

Modules

input

logic

model

Functions

puzzle_metadata